% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CovCoagEmaValidation.R
\docType{package}
\name{CovCoagEmaValidation}
\alias{CovCoagEmaValidation}
\title{CovCoagEmaValidation}
\description{
A package for validating prediction models using data in the OMOP CDM
}
